pyproject.toml
[tool.poetry]
name = "algorithmic-trader"
version = "0.1.0"
description = "Modular, cloud-native algorithmic trading platform for intraday & multi-bucket strategies"
authors = ["Algorithmic Trader Team"]
readme = "README.md"
packages = [
    {include = "packages"},
    {include = "apps"},
    {include = "ui"}
]

[tool.poetry.dependencies]
python = "^3.11"
# Data & Yahoo Finance
yfinance = "^0.2.28"
pandas = "^2.2.0"
numpy = "^1.26.0"
pyarrow = "^15.0.0"
# Pydantic for config & validation
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
# Alpaca trading
alpaca-py = "^0.20.0"
# Technical analysis
ta-lib = "^0.4.28"
# Streamlit UI
streamlit = "^1.31.0"
plotly = "^5.18.0"
st-aggrid = "^0.3.4"
# HTTP & async
httpx = "^0.26.0"
aiohttp = "^3.9.0"
# Database
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
# AWS
boto3 = "^1.34.0"
# Logging & monitoring
structlog = "^23.2.0"
prometheus-client = "^0.19.0"
# Scheduling & async
celery = "^5.3.0"
redis = "^5.0.0"
# Configuration
PyYAML = "^6.0.1"
python-dotenv = "^1.0.0"
# Date handling
python-dateutil = "^2.8.2"
pytz = "^2023.4"
# CLI
click = "^8.1.7"
rich = "^13.7.0"
# QuantConnect integration
lean = "^1.0.195"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
# Linting & formatting
ruff = "^0.1.15"
black = "^23.12.0"
isort = "^5.13.0"
mypy = "^1.8.0"
# Type stubs
types-PyYAML = "^6.0.12"
types-python-dateutil = "^2.8.19"
types-redis = "^4.6.0"
pandas-stubs = "^2.1.4"
# Jupyter for analysis
jupyter = "^1.0.0"
ipykernel = "^6.29.0"

[tool.poetry.scripts]
screener = "apps.screener.cli:main"
analyzer = "apps.analyzer.cli:main"
strategy = "apps.strategy.cli:main"
execution = "apps.execution.cli:main"
ui = "ui.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "yfinance.*",
    "alpaca.*",
    "ta.*",
    "streamlit.*",
    "plotly.*",
    "st_aggrid.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]

[tool.black]
line-length = 100
target-version = ['py311']
